+++ HolyMoly_Text_RPG/Goblin.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "Monster.h"
+
+class Goblin :public Monster {
+public:
+	Goblin(int level) : Monster(level) {
+		name = "Goblin";
+	}
+	Item* dropItem() override;
+};
