+++ HolyMoly_Text_RPG/Orc.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "Monster.h"
+
+class Orc :public Monster {
+public:
+	Orc(int level) : Monster(level) {
+		name = "Orc";
+	}
+	Item* dropItem() override;
+};
