+++ HolyMoly_Text_RPG/Troll.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "Monster.h"
+
+class Troll :public Monster {
+public:
+	Troll(int level) : Monster(level) {
+		name = "Troll";
+	}
+	Item* dropItem() override;
+};
